Change bnb_4bit_compute_dtype to torch.float16 for improved performance

+++ src/train.py
@@ -71,7 +71,7 @@ def main():
         load_in_4bit=True,
         bnb_4bit_use_double_quant=True,
         bnb_4bit_quant_type="nf4",
-        bnb_4bit_compute_dtype=torch.bfloat16
+        bnb_4bit_compute_dtype=torch.float16
     )
 
     model = AutoModelForCausalLM.from_pretrained(
